+++ test/compilationTests/gnosis/Tokens/Token.sol
@@ -8,16 +8,16 @@ contract Token {
     /*
      *  Events
      */
-    event Transfer(address indexed from, address indexed to, uint value);
-    event Approval(address indexed owner, address indexed spender, uint value);
+    event Transfer(address indexed from, address indexed to, uint256 value);
+    event Approval(address indexed owner, address indexed spender, uint256 value);
 
     /*
      *  Public functions
      */
-    function transfer(address to, uint value) public returns (bool);
-    function transferFrom(address from, address to, uint value) public returns (bool);
-    function approve(address spender, uint value) public returns (bool);
-    function balanceOf(address owner) public constant returns (uint);
-    function allowance(address owner, address spender) public constant returns (uint);
-    function totalSupply() public constant returns (uint);
+    function transfer(address to, uint256 value) public returns (bool);
+    function transferFrom(address from, address to, uint256 value) public returns (bool);
+    function approve(address spender, uint256 value) public returns (bool);
+    function balanceOf(address owner) public constant returns (uint256);
+    function allowance(address owner, address spender) public constant returns (uint256);
+    function totalSupply() public constant returns (uint256);
 }
